+++ test/graph.js
@@ -6,7 +6,7 @@ exports[ "lifecycle" ] = {
 		done();
 	},
 	test: function ( test ) {
-		test.expect( 28 );
+		test.expect( 35 );
 		test.equal( this.graph.get_node_value( 'f' ), undefined, "Should be `undefined`" );
 		test.equal( this.graph.set_node_value( 'f', 3 ), true, "Should be `true`" );
 		test.equal( this.graph.get_node_value( 'f' ), 3, "Should be `3`" );
@@ -30,11 +30,18 @@ exports[ "lifecycle" ] = {
 		test.equal( this.graph.get_edge_value( 'h', 'f' ), undefined, "Should be `undefined`" );
 		test.equal( this.graph.set_node_value( 'n', 0 ), true, "Should be `true`" );
 		test.equal( this.graph.get_node_value( 'n' ), 0, "Should be `0`" );
-		test.equal( this.graph.set_node_value( 'o' ), true, "Should be `true`" );
+		test.equal( this.graph.set_node_value( 'o', undefined ), true, "Should be `true`" );
 		test.equal( this.graph.get_node_value( 'o' ), null, "Should be `null`" );
 		test.equal( this.graph.add( 'n', 'o' ), true, "Should be `true`" );
 		test.equal( this.graph.set_edge_value( 'n', 'o', undefined ), true, "Should be `true`" );
 		test.equal( this.graph.get_edge_value( 'n', 'o' ), null, "Should be `null`" );
+		test.equal( this.graph.del_node( 'o' ), true, "Should be `true`" );
+		test.equal( this.graph.adjacent( 'n', 'o' ), false, "Should be `false`" );
+		test.equal( this.graph.neighbors( 'n' ).length, 0, "Should be `0`" );
+		test.equal( this.graph.get_edge_value( 'n', 'o' ), undefined, "Should be `undefined`" );
+		test.equal( this.graph.set_node_value( 'x', undefined ), true, "Should be `true`" );
+		test.equal( this.graph.get_node_value( 'x' ), null, "Should be `null`" );
+		test.equal( this.graph.del_node( 'z' ), false, "Should be `false`" );
 		test.done();
 	}
 };